backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
django-filter>=23.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=4.5.0
PyJWT>=2.8.0
django-extensions>=3.2.0
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.5.0
Pillow>=10.0.0
django-rest-auth>=0.9.5
django-allauth>=0.57.0
dj-rest-auth>=5.0.0
cryptography>=41.0.0